requirements.txt
requests==2.32.5
aiohttp==3.9.1
psycopg2-binary==2.9.10
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyJWT==2.10.1
google-generativeai>=0.3.1
python-dotenv==1.0.0